feat: removed chat prototype, added Subscription Hook
